Separate out concurrent tests and assert the concurrency.
