+++ core/server/src/test/java/alluxio/master/file/FileSystemMasterTest.java
@@ -17,7 +17,6 @@ import alluxio.Configuration;
 import alluxio.ConfigurationTestUtils;
 import alluxio.Constants;
 import alluxio.PropertyKey;
-import alluxio.collections.ConcurrentHashSet;
 import alluxio.exception.BlockInfoException;
 import alluxio.exception.DirectoryNotEmptyException;
 import alluxio.exception.ExceptionMessage;
@@ -40,7 +39,6 @@ import alluxio.master.file.options.SetAttributeOptions;
 import alluxio.master.journal.JournalFactory;
 import alluxio.security.GroupMappingServiceTestUtils;
 import alluxio.security.LoginUserTestUtils;
-import alluxio.security.authentication.AuthenticatedClientUser;
 import alluxio.thrift.Command;
 import alluxio.thrift.CommandType;
 import alluxio.thrift.FileSystemCommand;
@@ -55,7 +53,6 @@ import alluxio.wire.LoadMetadataType;
 import alluxio.wire.TtlAction;
 import alluxio.wire.WorkerNetAddress;
 
-import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
@@ -75,17 +72,12 @@ import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.CyclicBarrier;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * Unit tests for {@link FileSystemMaster}.
@@ -1233,374 +1225,5 @@ public final class FileSystemMasterTest {
     CompleteFileOptions options = CompleteFileOptions.defaults().setUfsLength(Constants.KB);
     mFileSystemMaster.completeFile(uri, options);
     return blockId;
-  }
-
-  /**
-   * Uses the integer suffix of a path to determine order. Paths without integer suffixes will be
-   * ordered last.
-   */
-  private class IntegerSuffixedPathComparator implements Comparator<FileInfo> {
-    @Override
-    public int compare(FileInfo o1, FileInfo o2) {
-      return extractIntegerSuffix(o1.getName()) - extractIntegerSuffix(o2.getName());
-    }
-
-    private int extractIntegerSuffix(String name) {
-      Pattern p = Pattern.compile("\\D*(\\d+$)");
-      Matcher m = p.matcher(name);
-      if (m.matches()) {
-        return Integer.parseInt(m.group(1));
-      } else {
-        return Integer.MAX_VALUE;
-      }
-    }
-  }
-
-  /**
-   * Tests concurrent renames within the root do not block on each other.
-   */
-  @Test
-  public void rootConcurrentRename() throws Exception {
-    final int numThreads = 100;
-    AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    AlluxioURI[] dsts = new AlluxioURI[numThreads];
-
-    for (int i = 0; i < numThreads; i++) {
-      srcs[i] = new AlluxioURI("/file" + i);
-      mFileSystemMaster.createFile(srcs[i], sNestedFileOptions);
-      dsts[i] = new AlluxioURI("/renamed" + i);
-    }
-
-    int errors = concurrentRename(srcs, dsts);
-
-    Assert.assertEquals("More than 0 errors: " + errors, 0, errors);
-    List<FileInfo> files =
-        mFileSystemMaster.listStatus(new AlluxioURI("/"), ListStatusOptions.defaults());
-    Collections.sort(files, new IntegerSuffixedPathComparator());
-    for (int i = 0; i < numThreads; i++) {
-      Assert.assertEquals(dsts[i].getName(), files.get(i).getName());
-    }
-    Assert.assertEquals(numThreads, files.size());
-  }
-
-  /**
-   * Tests concurrent renames within a folder do not block on each other.
-   */
-  @Test
-  public void folderConcurrentRename() throws Exception {
-    final int numThreads = 100;
-    AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    AlluxioURI[] dsts = new AlluxioURI[numThreads];
-
-    AlluxioURI dir = new AlluxioURI("/dir");
-
-    mFileSystemMaster.createDirectory(dir, CreateDirectoryOptions.defaults());
-
-    for (int i = 0; i < numThreads; i++) {
-      srcs[i] = dir.join("/file" + i);
-      mFileSystemMaster.createFile(srcs[i], sNestedFileOptions);
-      dsts[i] = dir.join("/renamed" + i);
-    }
-    int errors = concurrentRename(srcs, dsts);
-
-    Assert.assertEquals("More than 0 errors: " + errors, 0, errors);
-    List<FileInfo> files =
-        mFileSystemMaster.listStatus(dir, ListStatusOptions.defaults());
-    Collections.sort(files, new IntegerSuffixedPathComparator());
-    for (int i = 0; i < numThreads; i++) {
-      Assert.assertEquals(dsts[i].getName(), files.get(i).getName());
-    }
-    Assert.assertEquals(numThreads, files.size());
-  }
-
-  /**
-   * Tests that many threads concurrently renaming the same file will only succeed once.
-   */
-  @Test
-  public void sameFileConcurrentRename() throws Exception {
-    int numThreads = 10;
-    final AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    final AlluxioURI[] dsts = new AlluxioURI[numThreads];
-    for (int i = 0; i < numThreads; i++) {
-      srcs[i] = new AlluxioURI("/file");
-      dsts[i] = new AlluxioURI("/renamed" + i);
-    }
-
-    // Create the one source file
-    mFileSystemMaster.createFile(srcs[0], CreateFileOptions.defaults());
-
-    int errors = concurrentRename(srcs, dsts);
-
-    // We should get an error for all but 1 rename
-    Assert.assertEquals(numThreads - 1, errors);
-
-    List<FileInfo> files =
-        mFileSystemMaster.listStatus(new AlluxioURI("/"), ListStatusOptions.defaults());
-
-    // Only one renamed file should exist
-    Assert.assertEquals(1, files.size());
-    Assert.assertTrue(files.get(0).getName().startsWith("renamed"));
-  }
-
-  /**
-   * Tests that many threads concurrently renaming the same directory will only succeed once.
-   */
-  @Test
-  public void sameDirConcurrentRename() throws Exception {
-    int numThreads = 10;
-    final AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    final AlluxioURI[] dsts = new AlluxioURI[numThreads];
-    for (int i = 0; i < numThreads; i++) {
-      srcs[i] = new AlluxioURI("/dir");
-      dsts[i] = new AlluxioURI("/renamed" + i);
-    }
-
-    // Create the one source directory
-    mFileSystemMaster.createDirectory(srcs[0], CreateDirectoryOptions.defaults());
-    mFileSystemMaster.createFile(new AlluxioURI("/dir/file"), CreateFileOptions.defaults());
-
-    int errors = concurrentRename(srcs, dsts);
-
-    // We should get an error for all but 1 rename
-    Assert.assertEquals(numThreads - 1, errors);
-    // Only one renamed dir should exist
-    List<FileInfo> existingDirs =
-        mFileSystemMaster.listStatus(new AlluxioURI("/"), ListStatusOptions.defaults());
-    Assert.assertEquals(1, existingDirs.size());
-    Assert.assertTrue(existingDirs.get(0).getName().startsWith("renamed"));
-    // The directory should contain the file
-    List<FileInfo> dirChildren =
-        mFileSystemMaster.listStatus(new AlluxioURI(existingDirs.get(0).getPath()),
-            ListStatusOptions.defaults());
-    Assert.assertEquals(1, dirChildren.size());
-  }
-
-  /**
-   * Tests renaming files concurrently to the same destination will only succeed once.
-   */
-  @Test
-  public void sameDstConcurrentRename() throws Exception {
-    int numThreads = 10;
-    final AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    final AlluxioURI[] dsts = new AlluxioURI[numThreads];
-    for (int i = 0; i < numThreads; i++) {
-      srcs[i] = new AlluxioURI("/file" + i);
-      mFileSystemMaster.createFile(srcs[i], CreateFileOptions.defaults());
-      dsts[i] = new AlluxioURI("/renamed");
-    }
-
-    int errors = concurrentRename(srcs, dsts);
-
-    // We should get an error for all but 1 rename.
-    Assert.assertEquals(numThreads - 1, errors);
-
-    List<FileInfo> files =
-        mFileSystemMaster.listStatus(new AlluxioURI("/"), ListStatusOptions.defaults());
-    // Store file names in a set to ensure the names are all unique.
-    Set<String> renamedFiles = new HashSet<>();
-    Set<String> originalFiles = new HashSet<>();
-    for (FileInfo file : files) {
-      if (file.getName().startsWith("renamed")) {
-        renamedFiles.add(file.getName());
-      }
-      if (file.getName().startsWith("file")) {
-        originalFiles.add(file.getName());
-      }
-    }
-    // One renamed file should exist, and 9 original source files
-    Assert.assertEquals(numThreads, files.size());
-    Assert.assertEquals(1, renamedFiles.size());
-    Assert.assertEquals(numThreads - 1, originalFiles.size());
-  }
-
-  /**
-   * Tests renaming files concurrently from one directory to another succeeds.
-   */
-  @Test
-  public void twoDirConcurrentRename() throws Exception {
-    int numThreads = 10;
-    final AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    final AlluxioURI[] dsts = new AlluxioURI[numThreads];
-    AlluxioURI dir1 = new AlluxioURI("/dir1");
-    AlluxioURI dir2 = new AlluxioURI("/dir2");
-    mFileSystemMaster.createDirectory(dir1, CreateDirectoryOptions.defaults());
-    mFileSystemMaster.createDirectory(dir2, CreateDirectoryOptions.defaults());
-    for (int i = 0; i < numThreads; i++) {
-      srcs[i] = dir1.join("file" + i);
-      mFileSystemMaster.createFile(srcs[i], CreateFileOptions.defaults());
-      dsts[i] = dir2.join("renamed" + i);
-    }
-
-    int errors = concurrentRename(srcs, dsts);
-
-    // We should get no errors
-    Assert.assertEquals(0, errors);
-
-    List<FileInfo> dir1Files =
-        mFileSystemMaster.listStatus(dir1, ListStatusOptions.defaults());
-    List<FileInfo> dir2Files =
-        mFileSystemMaster.listStatus(dir2, ListStatusOptions.defaults());
-
-    Assert.assertEquals(0, dir1Files.size());
-    Assert.assertEquals(numThreads, dir2Files.size());
-
-    Collections.sort(dir2Files, new IntegerSuffixedPathComparator());
-    for (int i = 0; i < numThreads; i++) {
-      Assert.assertEquals(dsts[i].getName(), dir2Files.get(i).getName());
-    }
-  }
-
-  /**
-   * Tests renaming files concurrently from and to two directories succeeds.
-   */
-  @Test
-  public void acrossDirConcurrentRename() throws Exception {
-    int numThreads = 20;
-    final AlluxioURI[] srcs = new AlluxioURI[numThreads];
-    final AlluxioURI[] dsts = new AlluxioURI[numThreads];
-    AlluxioURI dir1 = new AlluxioURI("/dir1");
-    AlluxioURI dir2 = new AlluxioURI("/dir2");
-    mFileSystemMaster.createDirectory(dir1, CreateDirectoryOptions.defaults());
-    mFileSystemMaster.createDirectory(dir2, CreateDirectoryOptions.defaults());
-    for (int i = 0; i < numThreads; i++) {
-      // Dir1 has even files, dir2 has odd files.
-      if (i % 2 == 0) {
-        srcs[i] = dir1.join("file" + i);
-        dsts[i] = dir2.join("renamed" + i);
-      } else {
-        srcs[i] = dir2.join("file" + i);
-        dsts[i] = dir1.join("renamed" + i);
-      }
-      mFileSystemMaster.createFile(srcs[i], CreateFileOptions.defaults());
-    }
-
-    int errors = concurrentRename(srcs, dsts);
-
-    // We should get no errors.
-    Assert.assertEquals(0, errors);
-
-    List<FileInfo> dir1Files =
-        mFileSystemMaster.listStatus(dir1, ListStatusOptions.defaults());
-    List<FileInfo> dir2Files =
-        mFileSystemMaster.listStatus(dir2, ListStatusOptions.defaults());
-
-    Assert.assertEquals(numThreads / 2, dir1Files.size());
-    Assert.assertEquals(numThreads / 2, dir2Files.size());
-
-    Collections.sort(dir1Files, new IntegerSuffixedPathComparator());
-    for (int i = 1; i < numThreads; i += 2) {
-      Assert.assertEquals(dsts[i].getName(), dir1Files.get(i / 2).getName());
-    }
-
-    Collections.sort(dir2Files, new IntegerSuffixedPathComparator());
-    for (int i = 0; i < numThreads; i += 2) {
-      Assert.assertEquals(dsts[i].getName(), dir2Files.get(i / 2).getName());
-    }
-  }
-
-  /**
-   * Helper for renaming a list of paths concurrently. Assumes the srcs are already created and
-   * dsts do not exist.
-   *
-   * @param src list of source paths
-   * @param dst list of destination paths
-   * @return how many errors occurred
-   */
-  private int concurrentRename(final AlluxioURI[] src, final AlluxioURI[] dst)
-      throws Exception {
-    final int numFiles = src.length;
-    final CyclicBarrier barrier = new CyclicBarrier(numFiles);
-    List<Thread> threads = new ArrayList<>(numFiles);
-    // If there are exceptions, we will store them here.
-    final ConcurrentHashSet<Throwable> errors = new ConcurrentHashSet<>();
-    Thread.UncaughtExceptionHandler exceptionHandler = new Thread.UncaughtExceptionHandler() {
-      public void uncaughtException(Thread th, Throwable ex) {
-        errors.add(ex);
-      }
-    };
-    for (int i = 0; i < numFiles; i++) {
-      final int iteration = i;
-      Thread t = new Thread(new Runnable() {
-        @Override
-        public void run() {
-          try {
-            AuthenticatedClientUser.set(TEST_USER);
-            barrier.await();
-            mFileSystemMaster.rename(src[iteration], dst[iteration]);
-          } catch (Exception e) {
-            Throwables.propagate(e);
-          }
-        }
-      });
-      t.setUncaughtExceptionHandler(exceptionHandler);
-      threads.add(t);
-    }
-    Collections.shuffle(threads);
-    for (Thread t : threads) {
-      t.start();
-    }
-    for (Thread t : threads) {
-      t.join();
-    }
-    return errors.size();
-  }
-
-  /**
-   * Tests that getFileInfo (read operation) either returns the correct file info or fails if it
-   * has been renamed while the operation was waiting for the file lock.
-   */
-  @Test
-  public void consistentGetFileInfo() throws Exception {
-    final int iterations = 100;
-    final AlluxioURI file = new AlluxioURI("/file");
-    final AlluxioURI dst = new AlluxioURI("/dst");
-    final CyclicBarrier barrier = new CyclicBarrier(2);
-    // If there are exceptions, we will store them here.
-    final ConcurrentHashSet<Throwable> errors = new ConcurrentHashSet<>();
-    Thread.UncaughtExceptionHandler exceptionHandler = new Thread.UncaughtExceptionHandler() {
-      public void uncaughtException(Thread th, Throwable ex) {
-        errors.add(ex);
-      }
-    };
-    for (int i = 0; i < iterations; i++) {
-      mFileSystemMaster.createFile(file, sNestedFileOptions);
-      Thread renamer = new Thread(new Runnable() {
-        @Override
-        public void run() {
-          try {
-            AuthenticatedClientUser.set(TEST_USER);
-            barrier.await();
-            mFileSystemMaster.rename(file, dst);
-            mFileSystemMaster.delete(dst, true);
-          } catch (Exception e) {
-            Assert.fail(e.getMessage());
-          }
-        }
-      });
-      renamer.setUncaughtExceptionHandler(exceptionHandler);
-      Thread reader = new Thread(new Runnable() {
-        @Override
-        public void run() {
-          try {
-            AuthenticatedClientUser.set(TEST_USER);
-            barrier.await();
-            FileInfo info = mFileSystemMaster.getFileInfo(file);
-            // If the file info is successfully obtained, then the path should match
-            Assert.assertEquals(file.getName(), info.getName());
-          } catch (InvalidPathException | FileDoesNotExistException e) {
-            // InvalidPathException - if the file is renamed while the thread waits for the lock.
-            // FileDoesNotExistException - if the file is fully renamed before the getFileInfo call.
-          } catch (Exception e) {
-            Assert.fail(e.getMessage());
-          }
-        }
-      });
-      reader.setUncaughtExceptionHandler(exceptionHandler);
-      renamer.start();
-      reader.start();
-      renamer.join();
-      reader.join();
-      Assert.assertTrue("Errors detected: " + errors, errors.isEmpty());
-    }
   }
 }
